Implement reading from files
This leverages serde_any's support for determining a format from a
filename. If that fails, the file is slurped into memory and all
deserialization formats are tried (as if --from is omitted).

It might be nice to ensure that the file is not slurped into memory if
an explicit --from is provided. But this isn't a huge deal.