Preserve ordering of keys in maps
This change pulls in a feature branch proposed for serde-value, but not
merged due to concerns about the implementation. In particular, the
public API of serde-value is different and incompatible with upstream in
this setup (though that's not too harmful for a binary-only tool like
recompose).

Discussion about the PR seems to have fizzled out since February of this
year. I might have to look into the suggestions presented and try to
understand them.