Select JSON formatting based on output device
jyt will now pretty-print JSON if outputting it to a terminal, and
compact-print JSON if outputting anywhere else. This means there's less
garbage being written out in the event that a human being isn't looking
at the output directly.

This option is not configurable, and at least for the moment I don't
plan to make it configurable. jyt output can be further processed by a
tool like jq to override the default choice.