Buffer input and output
Want to add on to my previous commit message...

To be 100% clear, when I say "major performance issues related to using
serde_json for input parsing," I do NOT mean to imply that there is any
problem whatsoever with serde_json. The "issues" are entirely related to
MY OWN PERSONAL FAILURE to implement buffering on inputs.

The performance issue just happened to not manifest when using the YAML
input type and the serde_yaml parser, since it turns out when serde_yaml
deserializes from a reader it secretly buffers all input into memory
with read_to_end. That is a pretty significant departure from my
assumptions, I knew that some copying would have to be involved (as very
obviously stated in the docs, and of course to support key YAML features
like anchors) but I was an idiot and figured it would be more "granular"
or something. Seeing this it's clear that the YAML input should just
work like the TOML input and use a slice, either mapped or pre-read into
memory.