Update README
First, remove the TODO about nicer errors, as this was implemented in
jyt v0.4.0.

Second, remove the thing about "avoiding internal buffers," since this
is partly determined by the internal implementation of each serializer
and deserializer, which is out of my control. serde_yaml in particular
seems to do a lot of "buffering" when it serializes that I wasn't
previously aware of.