Enable LTO for release builds
There is a minimal but measurable impact on performance (about 5%
improvement) for very large inputs. More notably, though, this change
seems to roughly halve the binary size produced by the build (the
difference is less dramatic after stripping).

In any case it does not look like build times are significantly impacted
by this option, since jyt is so small anyways, so I don't think it hurts
to have this on.