Refactor output handling out of transcode_to
This is intended to make transcode_to safer to call more than once,
which will be required for multi-document support in a future version of
jyt.

Ordinarily I would have put this on a branch for that specific
changeset, but there are other changes I want to make around jyt's
output and I'd like to have them all in place before thinking about
multi-document stuff.