Always emit a newline after a JSON value
See RFC 8259 Section 2. The JSON grammar allows "insignificant
whitespace" (including both line feed and carriage return characters)
before or after the the serialized value in a JSON text, so it should
always be acceptable to emit this final newline whether we are
outputting to a terminal or somewhere else. This is convenient for a
number of reasons; most notably it should help with a future
implementation of multi-document support in jyt.