Remove TOML output support
Because TOML is so heavily specialized for writing configuration files,
and cannot reasonably be considered a generic data serialization format,
it presents a number of awkward special cases as an output format for
jyt. Lack of support for null values, ordering of non-table values
before tables, and the inability to delimit multiple documents within a
single stream are all unlike the other two formats supported by jyt.
While I have in some cases used jyt to transcode TOML files to JSON for
analysis by jq, I have not yet had a real use case for outputting
anything as TOML rather than JSON or YAML.

So, in order to focus on multi-document support for jyt v0.5, I'm going
to remove the one format that can't do multi-document output so I don't
have to think about it, at least in this first pass. If I ever actually
want TOML output in the future I can probably think of a good way to add
it back in at that time.