Implement streaming for multi-document JSON
I finally put two things together in order to figure out this
implementation: serde_json's deserializer can be used more than once to
deserialize multiple documents, and includes a function that errors out
if it detects that the deserializer is not at the end of the stream.
It's a slightly weird-looking implementation but it does seem to work on
everything I've tested with.