Flip relationship between input and output
This is the groundwork for my planned change to bring back TOML output
support in a manner that respects TOML's inability to support multiple
independent documents within a single file.

To do this cleanly, I want each output format to have the ability to
maintain state across transcoding calls if necessary, so that the TOML
output can know if it has already transcoded a document. My original
implementation of jyt created each new output serializer on demand in
each iteration of the input loop, so this plan flips things on its head
by requiring the output to be created first and then passed to the input
loop.