Auto-detect input formats based on parsing
With jyt v0.5's switch to slice-based input for all formats, it's much
easier to implement input format auto-detection based on parsing trials.
This is similar to what the earliest versions of jyt were able to do
with the serde_any crate, but with the new features like multi-document
support also in place.

Documentation has been fairly extensively updated to talk about the new
functionality.