Avoid string conversion on deserializer errors
This change requires that the error type of every deserializer satisfy
'static bounds, so that we can pass deserializer errors up to the main
method without converting them to strings and losing potentially useful
information in the process (e.g. we don't want error output on broken
pipes).

This effort won't *really* be complete until I deal with the fact that
serde_transcode effectively stringifies all of the errors it handles. I
don't have a great solution for that right now.