Initial implementation of msgpack format
This is a slightly rough but functional start on handling MessagePack as
a standard jyt format. A little while ago I was playing around with the
backup files from a HashiCorp Consul cluster, which apparently use a
MessagePack-based encoding. I'm sure there are other scenarios where
dumping or producing MessagePack from other formats could be helpful.

The biggest issue is that my current implementation is unable to borrow
data from the input slice during transcoding, and always copies out of
the input slice into new buffers when deserializing. Plus, as described
in comments the library will allocate and zero-fill these buffers using
the lengths specified in the MessagePack input stream, making it easy to
burn tons of CPU and memory on invalid inputs (e.g. c6 ff ff ff ff). I'd
rather be able to detect the length of each MessagePack value in the
stream so that I can pass slices of the right lengths to the
deserializer, but this might require some custom work on my part.