Remove nonsensical "value too large" error
We are reading a MessagePack value from a slice in memory. Its real size
cannot possibly exceed the maximum usize for the current platform, thus
there is absolutely no sense in maintaining the size as a u64 and
converting it "down." If the size is misreported in the input itself or
if a partial value is loaded into memory some reason, that is simply a
normal truncation error that we already take care of.

I am absolutely ashamed and embarrassed to have come up with such
obviously ridiculous logic in my initial implementation of this feature.

In less humiliating news, I've renamed the "UnexpectedEof" error to
"Truncated," which just seems nicer? I also added a specific unit test
for nonsensically large inputs.