Initial implementation of streaming JSON input
There's a lot going on here, and to be honest it was surprisingly hard
to get right.

The general idea is that there is a new Input abstraction, which
represents either a reader or a slice, but that also provides convenient
methods for getting a slice regardless of what the original input was.
If you skip format detection and your input format supports streaming,
you get streaming input, otherwise it behaves pretty much as before.

I think there's still some good work I could do on the API to help
prevent accidental misuse, i.e. pulling the reader out of a &mut Input.