Add streaming input support for MessagePack
This reintroduces the issue where the MessagePack decoder will eagerly
preallocate buffers even when it's arguably dangerous to do so. For now
I'm just documenting this in the help output. If I ever decide to be a
good person maybe I'll find a way to contribute something useful
upstream. It looks like Rust has std::io::Take as the analogue of
something like Go's io.LimitReader, which seems like it might be a
better model. In fact the implementation of read_to_end in Take is smart
enough to take advantage of the limit when reserving new space in the
buffer.