Add initial implementation of transcode::Value
This is the more generalized form of json::BorrowedValue, which in
theory should help us get around the fact that serde_transcode
stringifies all errors.

Right now we still get a "yaml-rust fmt error" when we break a pipe
while writing YAML, so it looks like we still have a way to go.