Actually ignore broken pipes when outputting YAML
Seems like yaml-rust just... always drops the underlying cause of any
I/O error. So even if we made it available and went down the chain of
causes we wouldn't see the broken pipe. Instead I guess we would have to
match on yaml-rust's error specifically?

We now detect broken pipes by trying to flush the output, which is much
more reliable.