Actually thoroughly discard values
I decided "why the heck not" and made the discarding serializer actually
go through all of the "internal" data of maps and sequences and whatnot,
so that if someone does try to call ser::Error::custom at any point
we'll find that out.

As part of this I decided to implement a *much* more interesting variant
of the discard serializer DSL macro thing, where it actually uses
English words instead of operators to indicate when a method does
something interesting. So that's neat.