Implement UTF-16 and UTF-32 support for YAML
This changeset introduces the ability to detect and re-encode UTF-16 and
UTF-32 YAML documents, as required by the YAML 1.2 specification.

Originally I attempted to do this with the encoding_rs crate, however
that turned out not to be a good fit for our needs. I'm sure it's great
for those who need to implement the full WHATWG Encoding Standard,
however for our purposes it was fairly large (around 10% of the *entire*
release binary) and did not support UTF-32. In contrast, this new
implementation handles all of the required encodings, yet only adds
about 10 KB to the size of the release build.

There is a small update to the auto detection logic to try MessagePack
before YAML, in case a valid MessagePack input would accidentally
trigger one of YAML's encoding detection rules. For example, a
single-element MessagePack array containing the integer 0 encodes as
0x91 0x00, which matches one of the cases for UTF-16-LE YAML input.