Don't panic and crash on transcode errors
I did, in fact, make a huge mistake in the new transcoder by assuming
that I could just return Ok in the middle of deserializing a valid
value, as I feared might potentially be the case. In particular
serde_yaml does NOT like this, and will panic if a visitor returns
without having visited an entire map. Why? Who knows.

Since serde_transcode just builds up a whole big chain of stringified
errors it kind of sidesteps the issue. We probably need to follow a
similar idea to the Transcoder, where we stash the error internally.
Hopefully I can actually do that.

For now, I just drain the remainder of a valid map or sequence on
serializer errors. Effectively this means we can't break out of an input
early, even when we know it no longer matters (but at least we're
deserializing into an IgnoredAny, so it should be pretty efficient).