Start cleaning up the new Visitor
We don't need a newtype like we do with the transcoder, since we can
implement this for &mut and avoid interior mutability.

We can also put most of the "store the error" junk behind a macro.

Finally, we can play with the macro syntax for the Value visitor methods
a little, particularly by removing the self parameter since the
transcoder no longer needs that feature for its visitors.