Simplify internal transcoder error messages
If we ever capture and print these out someday, this will look a bit
nicer.

One nice thing about the serde_transcode way of doing things is that it
can potentially capture some better detail from the deserializer on
serialization errors, such as line and column info. However we need to
find a good way to make that work.