Initial capturing of deserializer errors
The first time we drop a deserializer error in favor of a serializer
error, we now capture that error so that we can provide additional
context for the serializer error where appropriate.

This just gets the basic plumbing in place, I'll try to clean it up and
explain the concept in the docs later on.