Add notes about how MessageCell should not exist
I kind of threw this together when I realized that my error capturing
strategy was dropping useful context from the deserializer. Looking at
this again with the benefit of time away from the problem, I think I can
see the path to implementing this with true error capturing as we do for
the original error.