Upgrade clap to v3, replacing structopt
The new v3 of clap with built-in derive support has been released.

There are a number of noticeable behavior changes in v3 compared to
structopt + clap v2, which I've tried to work around or adjust in cases
where I prefer the existing behavior.

The first change is that all flags / options are now "unified" by
default instead of being split into sections. I've manually added my own
help and version flags and defined my own headers in order to bring back
that split. In fact, I was even able to reorganize the help and version
flags to the bottom of the help output, so that the jyt-specific
arguments and flags are next to each other instead of split apart by
that section. I think with the low number of options that jyt has, this
is actually easier to understand than the unified help message (which
does feel like a much better default for longer lists of options).

The second change is that the short "-h" flag no longer prints the full
help output. By default, it only prints the one-line summary at the top
of the output, which makes it impossible to discover that there is a
very detailed help output including the full list of available formats
and the features they support. I ended up writing my own short about
message that references the long "--help" option, to keep this somewhat
discoverable while adhering more closely to clap's common convention of
giving these flags separate meanings.

The last change is that invalid format names now include an explicit
message to "try --help", similar to what happens when you pass an
invalid flag or option. This might be a side effect of having to use the
Display implementation for clap::Error (as the "message" field has been
made private), however I think it's a good change.