First pass of unified state machine transcoder
This is an updated transcoder implementation that better reuses the core
"state machine" abstraction across all parts of the process, and that
replaces custom macro syntaxes with regular generic functions.

The Forwarder is still implementing its own state machine logic for the
moment, since it needs a Cell to properly implement Serialize. I'm going
to try having the standard state machine use a Cell internally so that
everything can use that single abstraction, as long as it doesn't have
bad side effects.