Use const generics to clean up msgpack size code
Previously we read MessagePack lengths of different sizes by having a
macro write out a bunch of trait impls that some other generic function
could resolve. As part of the general "de-macro-fication" of this
release I figured it was a good time to finally clean up this code a
bit, and also gain what's probably my first actual experience with const
generics.