Rework giant match for msgpack sizes
Instead of the size after the marker byte, the match arms now directly
yield the total size of the value indicated by the marker, length, etc.
This feels more intuitive to me, especially when it comes to arrays and
maps and I can slice the input at the same index the static length
implies.

Speaking of which, I've also pulled out the try_read_length… calls for
arrays and maps to happen before the total size calls, since they
implicitly validate that we can slice the input directly to the byte
following the length value. Cleaning up the optional get and its
associated error handling makes these cases a lot easier to read.