More accurately represent transcode errors
See the updated comments for details. I realized it's not strictly true
that the presence of a serializer error indicates that the serializer
was responsible for the transcode failure, because for nested structures
it's entirely possible that we generated the error ourselves.

I suppose I could try to make the serializer think that things were
successful and only run through the deserializer's error path, but it
might be weird for a failed transcode to "successfully" translate a
truncated form of the input. Might need some more thought on this.