Wait! The answer is obvious!
WE JUST NEED MORE OUT OF BAND DATA!

The first time we generate a synthetic error, we mark the machine to
indicate the true source. We propagate that mark up the call stack. If
we end up with a serializer error, but the mark indicates that the
initial error came from the deserializer, we hide the serialization
error. That brings us back to the enum-style error that we had before.

This is going to be *quite* interesting to implement.