Simplify and generalize machine structure
Rather than an enum that flips states, a Machine now contains two
entirely separate slots: one for the "parent" (de)serializer, and one
for the captured error. This model is more generalizable to other
captured states, such as the state indicating which side of the
transcode initially generated an error.