Initial implementation of error source plumbing
With this, we should be able to successfully differentiate between
errors actually generated by the serializer itself, and errors that we
generated within the serializer to safely halt a transcode. By default,
errors are assumed to have been triggered by the deserializer. If we
encounter an explicit serializer error, we will capture that in the
current machine and propagate it up. If we have a child machine, we will
always propagate its error state instead of marking our own.

Will put some documentation updates in place before merging this branch
for real, and exercise the logic a bit with some testing.