Add initial criterion benchmarks
I like the idea of having this in the project itself, instead of me
doing ad-hoc hyperfine runs every time I want to look at the potential
performance impact of big changes.

The benchmarks are based on the same data that I normally use for
testing, highly compressed for inline distribution in the Git repo. The
benchmarks will automatically translate the data into the format we want
to test with before starting.