Unify the small and large benchmarks
I'm not sure how Criterion.rs behaves when you have two wholly separate
benchmarks trying to use the same group names, so I figure it makes more
sense to do this. Then if I ever use iai or something in the future it's
easier to keep that separate.

I compressed the K8s input so it can use the same underlying loader
function without any special cases, not because I think it's actually
useful for storage.