Add general README for benchmarking
This has some more general info about how Criterion.rs works, and some
info about how the small and large benchmarks differ. Of course it also
still includes the MIT License for the original json-iterator data,
which is secretly the real point of this README existing in the first
place.