Idea for streaming auto-detection
We currently treat JSON and TOML as the "strict" cases at the very start
of the auto-detection logic. But when you think about it, MessagePack is
a lot stricter than JSON, especially with our requirements around the
types of starting values we allow.

The two strictest data formats also happen to be the two formats that
support streaming. Sounds like something to take advantage of!

I'm thinking that we could have some kind of reader that maintains a
prefix of whatever it has read so far, that we can reset to the
beginning to include both the read and unread content. We give that
reader to some specialized detection functions in the MessagePack and
JSON modules, respectively, and see if either one can successfully
transcode a single value. If not, we fall back to buffering and try TOML
and YAML, respectively. For buffering we do pretty much what we do now,
except we make the order of the tests match up with the reader ordering.
Of course, if YAML gains support for streaming in the future, we should
be able to adapt to that as well. We still need to do the TOML check
first, but maybe we restrict TOML to a certain input size and otherwise
decide that it's too large for us to care (even something like 2 - 4 MB
should be enough, who the heck is writing TOML files that big?).

I can't find anything in the standard library or on crates.io that does
exactly what I want out of the box, so I'm probably going to have to do
some legwork on this myself.