Add initial skeleton for new input module
This new module includes a "rewindable reader" implementation from an
experimental private branch, which will play a key role in the format
detection implementation for streaming inputs.

The general idea of the rewritten input handle is to support borrowing
the input through a guard that rewinds the reader when dropped, and
otherwise to only support consuming the handle to obtain the original
input. This is so different from the current approach of cloning handles
and tracking reference counts at runtime that I really think it's
necessary to start from scratch to do this right.