Fill out the new InputHandle implementation
It turns out that this really only works if the handle directly contains
a slice reference with the original lifetime, and not if it contains a
trait object with a bound on that lifetime. I think the idea is: sure,
the container _itself_ lives for that lifetime, but when you actually
try to deref, the reference you get back takes on the lifetime of the
container's &self reference, which is somehow bound to the lifetime of
the &self you borrowed it from. I have to admit I don't actually fully
understand it, but that's the intuition I've come up with for now.