Don't keep an unnecessary Cow in the input
Originally I had this idea that when a reader was fully buffered we
could turn the input source into an Input::Buffer. Instead, though, I
think it makes more sense to do that transformation in borrow_mut, and
just present the reader's capture buffer as if it were a slice input.