Try to look for EOFs in YAML and TOML detection
This is not an especially reliable strategy.

At this point, I think the right approach is to allow streaming
detection for JSON and MessagePack, but keep the full buffering approach
for YAML and TOML. Preventing xt from using unbounded memory on
arbitrary inputs is a problem to solve some other time.