Restore full buffering for inputs that need it
This is a closer match for xt's previous behavior. For now, I won't
worry about the fact that xt will use unbounded memory if it can't
detect a streaming input format immediately (to be fair, it can use
unbounded memory now if the value at the start of the input is super
large).