Fix a subset of pedantic clippy lints
Some of the suggestions are clearly false positives, and to be perfectly
honest I don't agree with all of them (e.g. omitting a closure in just
one transcode visitor function looks strange when that's not possible
for any of the other 15 or whatever). On the other hand, things like
semicolon_if_nothing_returned are things I normally try to look out for
myself so I'm glad to see there's an automated lint for them. I've added
explicit deny rules for everything I've changed here.