Switch back to exclusive range
I'm actually going to go ahead and disagree with clippy on this one.
Inclusive ranges are so rare in my experience compared to exclusive
ranges that I do a bit of a double take when I see one, and have to
think a lot more carefully about whether it actually does what I want.
Frankly, I think this form makes it a lot more obvious at a glance that
we're looking for a slice N bytes long offset by 1 from the start.