Reset source_eof if necessary
This is a bit of a moot point with respect to xt's usage of this reader,
since after the first EOF everything above `CapturedReader` will assume
the source is buffered and just expose the reader's slice while hiding
the fact that it came from a reader at all. But it's another place where
I should have done better the first time so it makes me feel better to
do it this way.