Fix partially consumed readers to not drop input
My first attempt to remove `Drop`-based reader rewinding introduced its
own issue: I forgot to also rewind the cursor when consuming a handle
and producing an input reader. This is now fixed, and is now covered by
a test case.

Basically, every time we expose a CaptureReader to anything outside of
this module, we have to make sure that it's rewound first. I'm not sure
if there's a better way to capture that in the code.