Remove rustfmt.toml; switch from 2 to 4 spaces
The previous 2-space indentation setting in rustfmt.toml was the result
of significant amounts of code being written in this style for reasons I
don't really understand (was I not applying rustfmt at all?).

Now that I am aware of the `--ignore-revs-file` option to `git blame`,
and GitHub's support for a `.git-blame-ignore-revs` file at the root of
the repository, I'd like to migrate over to the default and standardized
setting accepted by the Rust community.