Use standard behavior for broken pipes on Unix
I was completely ignorant of the backstory behind SIGPIPE in Rust, and I
suppose to some extent of SIGPIPE in general, and it never really
occurred to me that the program _should_ be killed by a signal and
therefore _should not_ just exit normally.

There are still some open concerns with this implementation, see
comments for details.