Use hard tabs for horizontal indentation
See comments in rustfmt.toml for references regarding the impact of this
choice on developers with visual disabilities.

As a developer without visual disabilities, this is a perspective I have
long failed to consider. To those who have taken the time and energy to
call attention to the harm that I have caused, thank you.