Reduce use of spaces for alignment in msgpack tests
One of the rustfmt.toml links points to a comment from MarcoZehe on the
Prettier repository that discusses how refreshable braille displays
treat spaces in code. rustfmt's original formatting of these comments,
especially around the fixext section, had a ton of wasted space that
seems like it could get annoying to read. Really, this version seems
easier to understand even to me as a programmer looking at a monitor.