Introduce a reusable translator type
As part of the planned support for multiple input files, this change
introduces a new "Translator" type that makes it possible to reuse a
single output (and its associated state) for multiple independent
inputs. The original xt::translate function is now just a basic wrapper
for this logic.

I've hand-written a lot of static dispatch code to support the
translator storing one of 4 possible output types. I tried doing this
with the enum_dispatch crate, but it seems like that crate can only
implement traits on the enum type, not on a &mut reference to the enum
type.