Refactor CLI to better support multiple inputs
Rather than methods on the Cli struct itself, a lot of the logic for
dealing with files is now factored into functions that just take file
paths. This should make it a lot easier to transform the Option for the
filename into a Vec (or however clap handles multiple arguments).