Support multiple input files in the CLI
This is pretty cool!

While making this change, I learned of another apparent benefit of the
serde-yaml rewrite: the serializer now properly reports I/O errors in
its error chain, which means we no longer have to flush after every
document to see if the output pipe is broken.

This is making me wish I had some more proper tests on the actual CLI,
and not just the translation library.