Raise SIGPIPE manually on broken pipe errors
I'm a big fan of how Go handles SIGPIPE errors: if the error happens on
stdout or stderr, die with SIGPIPE to work like a typical command line
utility; otherwise, ignore the signal to avoid dying on closed network
connections.

When I went to study how this was implemented, I assumed that the Go
runtime had some special way to link a received SIGPIPE to a write on a
particular file descriptor (which in retrospect seems a bit ridiculous).
In reality, the runtime basically ignores SIGPIPE (at least as long as
no C code installs a handler for it), and when the os.File.Write method
encounters an EPIPE it calls into a special runtime method that fiddles
with the necessary handler settings and manually raises SIGPIPE.

Honestly, it never even crossed my mind that xt could raise its own
signals, even though in theory it's something I know Unix programs are
capable of. I think this is a much cleaner way to achieve the goal of
looking like a "normal" Unix utility without maintaining and (if I'm
being honest, not actually) testing separate Unix and non-Unix error
handling paths.