Refactor transcoder to use more newtypes
With this, I'm replacing the pattern of implementing a whole bunch of
things directly on a `Transcoder` struct with the newtype pattern that I
had to use for `KeySeed` and `ValueSeed`. This feels a lot easier to
understand than when you have to look carefully at the generics on each
impl block to figure out all the ways a single type can actually be
used.