Comment and naming tweaks
I'm trying to remove references to specific types in my internal
documentation wherever possible, since this was one of the more annoying
things about the recent transcoder work, and is frankly a best practice
that I need to get a better handle on.

In other news, I also tweaked some of the naming and docs around the
"give me a prefix of the input" logic in a way that I think is a lot
more intuitive. I don't know that it's _perfect_, but I don't feel as
awkward about it as I did before.