Remove note about re-assessing LTO
The actual performance and size characteristics of the various possible
settings vary pretty widely, but overall it seems like full LTO with 1
codegen unit is still a winner once you actually look into it. At the
very least, I think it makes sense as the default for publishing to
Crates.io, where rebuilds are going to be relatively rare.