Remove TODOs about persisting UTF-8 check results
After experimenting with this *significantly* more than I should have,
I've resoundingly concluded that there is no value whatsoever in sharing
the results of UTF-8 validity checks between different stages of the
translation process. The overhead of the check is so little compared to
actual parsing that it's practically statistical noise.

The one notable finding from my work was that an upfront UTF-8 check for
JSON slice inputs made a significant difference in JSON to MessagePack
translation performance, so I've gone ahead and implemented that. But
even when sharing that result between auto-detection and translation for
a JSON input without a .json file extension, whatever improvement we get
is effectively statistical noise.