Update v0.11.0 changelog to call out escaping bugs
Some time after the serde_yaml upgrade, I found that a test document I'd
serialized with an old version of serde_yaml was no longer readable by
the new version due to the inclusion of an unescaped U+0081 control
character in a string. Of course, since libyaml is (often) the basis for
YAML support in Python, Ruby, and any number of other popular tools,
this is a pretty notable issue that I think deserves an explicit callout
as a bug fix.