Release v0.12.2
I did go back and forth a bit on whether this was worth a full MINOR
bump. In the end I see both of these changes as polishing the major
feature of the v0.12.0 release, and not as changes that I would expect
to break anything or that I would be willing to market as something
entirely new.