Introduce the new xt::yaml::chunker module
This module, originally developed as part of an external prototype
project, provides the ability to chunk YAML text streams into their
constituent documents so that serde_yaml can parse them individually.

As the documentation on the module makes clear, this is "the evil way"
that I've previously discussed in the comments of xt::yaml::transcode,
and I sincerely hope I can find the time and energy to take what I've
learned from developing this module and do this "the righteous way."

Like with the encoding module, this is a minimal integration with a lot
of dead code, and the only changes from the prototype are some Clippy
fixups and pub(super) additions. My first goal is to get streaming
translation to work with YAML explicitly specified as the source format,
since streaming YAML auto-detection runs into the issue where some TOML
documents parse successfully as plain flow scalars. I'll probably need
to do something like MessagePack and limit detection to maps and arrays,
but this is going to require some interesting work and thinking.