Introduce initial support for streaming YAML input
Thanks to the new streaming text encoder and YAML chunker, we've finally
achieved the first working implementation of streaming YAML input!

Besides the awful hackishness of the implementation that I've written
about at length in the comments, the main limitation of this first pass
is that YAML streaming only works when format detection is bypassed.

My current thinking is that instead of returning a String, the chunker
can return a richer "Chunk" that includes both the string content and a
marker indicating whether the root of the document is a scalar,
sequence, or mapping. It should be pretty easy for the format detector
to take advantage of that, I just hope it doesn't complicate the chunker
too much more. The other possibility would be to parse the first chunk
with serde_yaml and somehow have it tell us the root type, but that
would probably require a custom Deserialize impl that I don't feel like
writing.