Introduce the new xt::yaml::encoding module
This module, originally developed as part of an external prototype
project, provides streaming text re-encoding support for YAML 1.2
streams. This is an important first step toward full support for
streaming YAML in xt, since even if serde_yaml eventually gains support
for streaming it is not guaranteed to gain full support for the set of
text encodings that xt handles.

This is a very minimal integration with a lot of dead code. The only
changes from the yaml-split prototype were to handle a couple of Clippy
lints that I'd enabled around potential truncation, and to change pub to
pub(super). My first goal is to replace xt's current YAML re-encoding
functions with the new streaming transcoder, even if the performance of
the reader-based transcode is a bit lower than the performance of the
slice-based transcode.