Update commentary around format detection
With the introduction of streaming YAML input support, we now need to
attempt YAML detection before TOML detection in order to avoid TOML's
attempts to buffer the full input. Because of this, I've re-worked some
of the explanation around how format detection works, and moved some of
the specific commentary around YAML scalars into the yaml module
(similar to what I did for MessagePack).