Fix mis-detection of TOML documents from readers
I've written a lot about the problem of a YAML parser accepting TOML
documents that don't start with tables as scalars. Now, I've run into
basically the opposite problem with the first pass of the new format
detection logic: the marker for a TOML table looks like a valid YAML
flow sequence!

In other words, when a YAML parser sees this document:

  [test]
  stuff = true

...it will happily parse that first line as a document containing the
sequence ["test"], and then fail once it gets to the second line. Of
course, since we only look at the first document in a YAML stream, and
since we emit that document as soon as the parser emits a document end
event, we don't get far enough to see the problem.

This is a hacky fix that gets us a bit closer to the desired behavior.
Now that I think about it, the real answer is probably to complete the
chunking of the document on the document end event, but to refuse to
emit that document until we reach a document start or stream end. That
is, we should modify the chunker instead of getting too weird with the
detection logic, since I don't want it to turn into some spaghetti mess.