Actually fix mis-detection of TOML documents from readers
This change updates the YAML chunker to avoid emitting a document until
it sees the start of the next document or the end of the YAML stream.
This ensures that if the prefix of an arbitrary input looks like YAML,
we won't accidentally assume that the entire input is YAML just before
getting to the part where the parser fails.