Enable strict lints on unsafe blocks and SAFETY comments
It sounds like there's a decent chance that `unsafe_op_in_unsafe_fn`
will become default-warn or default-deny in Rust 2024. While I see the
potential for noise, I like the idea in general and I'm interested in
both getting ahead of the curve and applying this for its own merits.

I've also decided to try Clippy's lint for SAFETY comments on all unsafe
blocks, since I think I do try to be *somewhat* decent about this, and I
think that the big loud warning will be a positive influence to keep me
on that track.